requirements.txt
orjson==3.10.4
zstandard==0.21.0
bifixer==0.8.12
iso639-lang==2.2.3
marisa-trie==1.2.0
docscorer @ git+https://github.com/pablop16n/web-docs-scorer@a469233
pii-manager @ git+https://github.com/mmanteli/multilingual-PII-tool@db59603
